requirements.txt
absl-py==0.11.0
aiofiles==0.8.0
allure-pytest==2.9.45
allure-python-commons==2.9.45
altgraph==0.17
amply==0.1.4
asgiref==3.5.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.5
beautifulsoup4==4.9.3
blinker==1.4
Brotli==1.0.9
bs4==0.0.1
cachetools==4.2.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.1.3
colorama==0.4.4
colorlog==6.6.0
comtypes==1.1.8
configparser==5.0.1
crypto==1.4.1
cryptography==36.0.0
cycler==0.10.0
Deprecated==1.2.13
DingtalkChatbot==1.5.3
Django==3.1.7
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-filter==2.4.0
djangorestframework==3.12.2
docutils==0.16
et-xmlfile==1.1.0
execnet==1.9.0
Faker==9.8.3
Flask==2.0.3
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h11==0.13.0
h2==4.1.0
h5py==2.10.0
hpack==4.0.0
httptools==0.4.0
hyperframe==6.0.1
idna==2.10
iniconfig==1.1.1
itchat==1.3.10
itsdangerous==2.1.2
jdcal==1.4.1
Jinja2==3.1.2
jsonpath==0.82
kaitaistruct==0.9
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
ldap3==2.9.1
loguru==0.6.0
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.4.0rc1
msgpack==1.0.3
multidict==6.0.2
mysqlclient==2.0.3
Naked==0.1.31
Nuitka==0.6.10.5
numpy==1.19.5
oauthlib==3.1.0
openpyxl==3.0.9
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.1.5
passlib==1.7.4
pefile==2019.4.18
Pillow==8.2.0
pluggy==1.0.0
protobuf==3.19.4
psutil==5.8.0
publicsuffix2==2.20191221
PuLP==2.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.21
pydantic==1.8.2
pyDes==2.0.1
pydivert==2.1.0
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.11
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyperclip==1.8.2
pypiwin32==223
pypng==0.0.21
PyQRCode==1.2.1
pytest==7.1.2
pytest-dependency==0.5.1
pytest-forked==1.3.0
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-order==1.0.1
pytest-ordering==0.6
pytest-xdist==2.4.0
python-dateutil==2.8.2
pytz==2020.5
pywin32==304
pywin32-ctypes==0.2.0
pywinauto==0.6.8
PyYAML==5.4.1
redis==4.3.4
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sanic==22.3.1
sanic-routing==22.3.0
scipy==1.6.1
seaborn==0.11.1
selenium==3.141.0
sentry-sdk==1.0.0
shellescape==3.8.1
shutils==0.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==6.1
trio==0.17.0
trio-websocket==0.9.0
typing-extensions==3.7.4.3
urllib3==1.26.7
urwid==2.1.2
websockets==10.3
Werkzeug==2.1.2
win32-setctime==1.1.0
wrapt==1.12.1
wsproto==1.1.0
xlrd==2.0.1
xlutils==2.0.0
xlwings==0.27.7
xlwt==1.3.0
zstandard==0.17.0
